# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Browser automation daemon — navigate, interact, and extract from a live headless browser. One browser per port.",
        "title": "webrudder API",
        "contact": {},
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/back": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interact"
                ],
                "summary": "Navigate back",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.GotoResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrResp"
                        }
                    }
                }
            }
        },
        "/batch": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interact"
                ],
                "summary": "Run multiple actions in one request",
                "parameters": [
                    {
                        "description": "ordered actions",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.BatchReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.BatchResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrResp"
                        }
                    }
                }
            }
        },
        "/check": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interact"
                ],
                "summary": "Set a checkbox/radio to checked/unchecked",
                "parameters": [
                    {
                        "description": "ref and desired checked state",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.CheckReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.OKResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrResp"
                        }
                    }
                }
            }
        },
        "/click": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interact"
                ],
                "summary": "Click an element by ref",
                "parameters": [
                    {
                        "description": "element ref",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.ClickReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.ClickResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrResp"
                        }
                    }
                }
            }
        },
        "/download": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interact"
                ],
                "summary": "Click an element, wait for the download, return the saved path",
                "parameters": [
                    {
                        "description": "ref and optional dir",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.DownloadReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.DownloadResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrResp"
                        }
                    }
                }
            }
        },
        "/fill": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interact"
                ],
                "summary": "Type text into an input/textarea by ref",
                "parameters": [
                    {
                        "description": "ref and text",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.FillReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.OKResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrResp"
                        }
                    }
                }
            }
        },
        "/forward": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interact"
                ],
                "summary": "Navigate forward",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.GotoResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrResp"
                        }
                    }
                }
            }
        },
        "/goto": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interact"
                ],
                "summary": "Navigate to an absolute URL (explicit jump)",
                "parameters": [
                    {
                        "description": "target url",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.GotoReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.GotoResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrResp"
                        }
                    }
                }
            }
        },
        "/hover": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interact"
                ],
                "summary": "Hover an element by ref",
                "parameters": [
                    {
                        "description": "ref",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.HoverReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.OKResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrResp"
                        }
                    }
                }
            }
        },
        "/html": {
            "get": {
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "read"
                ],
                "summary": "Raw outer HTML (full page, or a single element via ?ref)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "element ref for outerHTML; omit for the full page",
                        "name": "ref",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrResp"
                        }
                    }
                }
            }
        },
        "/press": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interact"
                ],
                "summary": "Press a key or chord (Enter, Tab, ArrowDown, Control+a)",
                "parameters": [
                    {
                        "description": "key",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.PressReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.OKResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrResp"
                        }
                    }
                }
            }
        },
        "/read": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "read"
                ],
                "summary": "Extract page URL, title, and visible text",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.ReadResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrResp"
                        }
                    }
                }
            }
        },
        "/reload": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interact"
                ],
                "summary": "Reload the current page",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.GotoResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "server.BatchReq": {
            "type": "object",
            "properties": {
                "actions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/browser.Action"
                    }
                }
            }
        },
        "server.BatchResp": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": {}
                    }
                }
            }
        },
...
        "server.ErrResp": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "ref \"e9\" not found — re-scan the page"
                },
                "ok": {
                    "type": "boolean"
                }
            }
        },
...
        "server.GotoResp": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "server.OKResp": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean"
                }
            }
        },
...
    }
}